+++ src/utils.js
@@ -1,4 +1,21 @@
+import React from 'react';
 import {Iterable} from 'immutable';
 
 export const ensureNativeJSValue = state => Iterable.isIterable(state) ? state.toJS() : state;
+
+export function toJS(WrappedComponent) {
+  function ToJSWrapper(wrappedComponentProps) {
+    const propsJS = Object.entries(wrappedComponentProps)
+      .reduce((newProps, [key, value]) => ({
+        ...newProps,
+        [key]: ensureNativeJSValue(value)
+      }), {});
+
+    return <WrappedComponent {...propsJS} />;
+  }
+
+  ToJSWrapper.displayName = `toJS(${WrappedComponent.displayName || WrappedComponent.name || 'Component'})`;
+
+  return ToJSWrapper;
+}
